+++ netcp_test.go
@@ -31,9 +31,9 @@ import (
 const RUN_CLIENT_TEST                bool = false
 
 /* Configuration */
-const CONTROLLER_DOMAIN                string = "127.0.0.1"
+const CONTROLLER_DOMAIN              string = "127.0.0.1"
 const CONTROLLER_PATH                string = "/gate.php"
-const CONTROLLER_PORT                int = 80
+const CONTROLLER_PORT                int16 = 80
 
 func TestMainChannel(t *testing.T) {
     /*
@@ -41,9 +41,9 @@ func TestMainChannel(t *testing.T) {
      */
     D("Building the server processor")
     D("Starting netcp service on [TCP] port: " + string(CONTROLLER_PORT))
-    service, err := CreateNetCPServer(    CONTROLLER_PATH, /* /gate.php */
+    service, err := CreateNetCPServer(   CONTROLLER_PATH, /* /gate.php */
                                          CONTROLLER_PORT, /* 80 */
-                                            0)
+                                         0)
     if err != nil || service == nil {
         D(err.Error())
         T("Cannot start netcp service")
